requirements.txt
mss
numpy
opencv-python
PyAutoGUI
PyYAML
pygetwindow